Move CircuitBreaker to internal
`CircuitBreaker` API is not stable yet.
