+++ src/main/java/com/atlassian/db/replica/internal/circuitbreaker/CircuitBreaker.java
@@ -1,0 +1,8 @@
+package com.atlassian.db.replica.internal.circuitbreaker;
+
+
+public interface CircuitBreaker {
+    BreakerState getState();
+
+    void handle(Throwable throwable);
+}
