+++ src/main/java/com/atlassian/db/replica/internal/circuitbreaker/ClosedBreaker.java
@@ -1,8 +1,5 @@
 package com.atlassian.db.replica.internal.circuitbreaker;
 
-import com.atlassian.db.replica.api.circuitbreaker.BreakerState;
-import com.atlassian.db.replica.spi.circuitbreaker.CircuitBreaker;
-
 public class ClosedBreaker implements CircuitBreaker {
     @Override
     public BreakerState getState() {
